Create PARSE-HTML.JS
The given code is a mixture of HTML, JavaScript and a third-party JavaScript library called "Clipboard.js".

The JavaScript code contains two functions named convert() and cdClear(). The convert() function takes the content of an HTML element with id="somewhere", replaces certain characters with their corresponding HTML entities, and then sets the new value as the content of the same HTML element. The function also makes two HTML elements with IDs "button-link" and "btn_clear" visible by setting their CSS display property to "inline-block".

The cdClear() function clears the content of the HTML element with id="somewhere" and hides an HTML element with id="btnInfo" and another HTML element with id="button-link" by setting their CSS display property to "none".

The third-party JavaScript library "Clipboard.js" is also included in the code, which allows copying text to the clipboard. However, it seems that no code is using this library in the given code snippet.